Refactor install to use a function table
